requirements.txt
pandas
numpy
yfinance>=0.2.40
ta
requests
python-dotenv
APScheduler
nltk
tqdm
